server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Petr Fialka <petr.fialka95@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
